commit for working version of jwt auth

+++ workout-webapp/frontend/src/components/WorkoutDetails.js
@@ -1,35 +1,38 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
-import { Link, useParams } from 'react-router-dom';
+import { useAuthContext } from '../hooks/useAuthContext'
+
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext()
-  const { id } = useParams();
+  const { user } = useAuthContext()
 
-  const handleDeleteClick = async () => {
+  const handleClick = async () => {
+    if (!user) {
+      return
+    }
+
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
 
     if (response.ok) {
       dispatch({type: 'DELETE_WORKOUT', payload: json})
     }
-  };
+  }
 
   return (
     <div className="workout-details">
-      { id === undefined && 
-        <Link to={`/workouts/${workout._id}`}><h4>{workout.title}</h4></Link>
-      }
-      { id != undefined && 
-         <h4>{workout.title}</h4>
-      }
+      <h4>{workout.title}</h4>
       <p><strong>Load (kg): </strong>{workout.load}</p>
-      <p><strong>Number of reps: </strong>{workout.reps}</p>
+      <p><strong>Reps: </strong>{workout.reps}</p>
       <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
-      <span className="material-symbols-outlined" onClick={handleDeleteClick}>delete</span>
+      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
     </div>
   )
 }
